Extract logic in FileBrowser useEffect into utility function
